// apps/user-onboarding/src/onBoard/customerInfo/typings/request.ts
export type SupplementalInformation = {
  transactionId: string;
  currentAddress: string;
  newCurrentAddress: string;
  newProvinceCode: string;
  newDistrictCode: string;
  newCommuneCode: string;
  newDetailAddress: string;
  mobilePhone: string;
  newMobilePhone: string;
  homePhone: string;
  newHomePhone: string;
  email: string;
  newEmail: string;
  currentOccupation: string;
  newCurrentOccupation: string;
  otherOccupationInfo: string;
  jobTitle: string;
  newJobTitle: string;
  otherJobTitleInfo: string;
  otherCurrentOccupation: string;
  otherJobTitle: string;
};